Add files via upload
Agro-Farm web application is the farmer system where they can plan, monitor and analyze the activity of the farmers production system. It manages farmer operation with one system and organizes data in one place. It helps smart farmers become even smarter. This creates in partnership with growers and buyers. It inspire farmer to produce and buyers to consume fresh goods. 
Agro-Farm Web application will make better connection among Farmers and Buyers ensure quality food. Standardize and increase efficiency of agro culture process .The name Agro-Farm Web application indicates Intelligent Agriculture. Agro-Farm Web application is a model farmer management website application. This site helps the farmers to sell their agricultural produce online and suggests best -in-practice farming processes. Hence, providing a wider market and helping them to not restrict themselves to the local market. This enables wholesalers and retailers to expand their business. 

Purpose of the project
The purpose of developing web application Agro-Farm Web app is to provide information to farmers who want to visit , sell  various products and easy to connectivity between buyer and markets etc.
Proposed of the Systems 
Agro-Farm Web application is farmer management website application which helps farmers to give best-practice farming processes. It helps farmers to improve their productivity and profitability. It enables farmers to sell their productions through online and farmers can purchase tools and seeds directly from seller.
TECHNOLOGY USED:
HTML
CSS
MySql
PHP
JavaScript
J-Query
XAMPP Server
Bootstrap
Visual Studio Code (Editor)

CONCLUSION
This Project will thus pave the way for an efficient means to carry out the buying and selling of the products. Farmers will earn money as per the work they have done and will not suffer losses. This system is proposed to replace the existing system where the farmer has to suffer between the manufacturers and the traders. Also the main advantage of this project is that it uses Information Technology. The User only needs basic products like a Computer and an internet connection.
Typically, the farmer agrees to provide established quantities of a specific agricultural product, meeting the quality standards and delivery schedule set by the purchaser. In turn, the buyer commits to purchase the product
This system is proposed to replace the existing system where the farmer has to suffer between the manufacturers and the traders. Also the main advantage of this project is that it uses Information Technology. The User only needs basic products like a Computer and an internet connection. 

Achievements Expected Through The Project
	Buyer can directly buy goods to farmer.
	Computing to existing system, it performs at a faster pace.
	Minimizing human efforts and cost efficient database.
	It simplifies the management process in travelling.
	The system provide greater processing speed consistency.
	Time taken for overall process is reduced.
	Ease of operation.